refactor(ui): use button composition for neutral actions
