+++ docs/styles.md
@@ -177,28 +177,30 @@ The `running-indicator`, `running-indicator-center`, and
 `running-indicator-ripple` recipes have been removed; their keyframes remain,
 since keyframes are not class selectors.
 
-### Neutral control surface
+### Neutral button and select variants
 
 Use `Button variant="neutral"` for neutral actions and
 `SelectTrigger variant="neutral"` for neutral select controls. Each component
 owns its utilities; their public API does not export class strings.
-`NeutralControl` from `@okouai/ui` preserves the neutral surface on native links
-and triggers. It renders a button by default; use `render={<Link ... />}` to
-retain a link's navigation semantics. It forwards refs and event handlers to
-that element without adding a wrapper.
+Use `Button asChild variant="neutral"` around a router `Link` for navigation
+styled as a button, and compose `Button` with `DialogTrigger` for dialog
+actions. The existing components own the interaction contract; `neutral` is
+only a visual variant. Link composition preserves the native anchor, ref,
+and navigation behavior without adding a wrapper.
 
 The components compose `border border-control-border bg-control-surface
 text-foreground [&:hover]:bg-state-hover-overlay` internally. This is the treatment the
 settings-select batch established, extended with the border and foreground the
 retired `okou-btn-morandi` selector owned. Language, timezone, and voice-input
 settings all use the select variant. Dimensions, padding, and radius remain
-with the existing control and caller; the appearance does not impose button
-layout on links or select triggers.
+with the existing component and caller.
 
 The neutral button adds the existing outline interaction fills
 (`hover:bg-state-hover active:bg-state-pressed`) to the shared surface, keeping
-the hover overlay above those fills. Select triggers and `NeutralControl` use
-the opaque surface and hover overlay without adding button interaction fills.
+the hover overlay above those fills. Select triggers retain the opaque surface
+and hover overlay. The existing recovery links and Add automation trigger
+preserve that same interaction treatment with `hover:bg-control-surface
+active:bg-control-surface` on their `Button` instances.
 
 Preserve consumer-specific interaction colors when extracting shared styles.
 The official workflow Configure button, for example, retains its existing
