+++ turbo/packages/ui/src/index.ts
@@ -20,10 +20,6 @@ export {
 } from "./components/ui/card";
 export { Badge, type BadgeProps } from "./components/ui/badge";
 export { IconButton, type IconButtonProps } from "./components/ui/icon-button";
-export {
-  NeutralControl,
-  type NeutralControlProps,
-} from "./components/ui/neutral-control";
 export { Checkbox } from "./components/ui/checkbox";
 export {
   ToggleButton,
